Move mock_connection and connection tests to `src` folder.
This means that we don't need to make public a lot of types that should be used only internally.
This moves the `MockConnection` logic into `src`, under the `testing` module.
