+++ redis/Cargo.toml
@@ -162,10 +162,6 @@ required-features = ["json", "serde/derive"]
 name = "test_cluster_async"
 required-features = ["cluster-async"]
 
-[[test]]
-name = "test_async_cluster_connections_logic"
-required-features = ["cluster-async"]
-
 [[bench]]
 name = "bench_basic"
 harness = false
